+++ packages/evmApi/src/resolvers/balance/getNativeBalance.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { Camelize } from '@moralisweb3/core';
-import { EvmAddressish, EvmChainish, EvmNative, EvmAddress } from '@moralisweb3/common-evm-utils';
+import { EvmAddressish, EvmChainish, EvmNative, EvmAddress } from '@moralisweb3/evm-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';