+++ packages/evmApi/src/resolvers/defi/getPairAddress.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { Camelize } from '@moralisweb3/core';
-import { Erc20Token, EvmAddress, EvmAddressish, EvmChainish } from '@moralisweb3/common-evm-utils';
+import { Erc20Token, EvmAddress, EvmAddressish, EvmChainish } from '@moralisweb3/evm-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';