+++ packages/evmApi/src/resolvers/token/getTokenAllowance.ts
@@ -1,5 +1,5 @@
 import { BigNumber, Camelize } from '@moralisweb3/core';
-import { EvmAddress, EvmAddressish, EvmChainish } from '@moralisweb3/common-evm-utils';
+import { EvmAddress, EvmAddressish, EvmChainish } from '@moralisweb3/evm-utils';
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';