+++ packages/evmApi/src/resolvers/token/getTokenMetadata.ts
@@ -1,7 +1,7 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 
 import { Camelize, toCamelCase } from '@moralisweb3/core';
-import { Erc20Token, EvmChainish, EvmAddressish, EvmAddress } from '@moralisweb3/common-evm-utils';
+import { Erc20Token, EvmChainish, EvmAddressish, EvmAddress } from '@moralisweb3/evm-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';