+++ packages/evmApi/src/resolvers/token/getTokenPrice.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { Camelize, toCamelCase } from '@moralisweb3/core';
-import { EvmNative, EvmAddress, EvmChainish, EvmAddressish } from '@moralisweb3/common-evm-utils';
+import { EvmNative, EvmAddress, EvmChainish, EvmAddressish } from '@moralisweb3/evm-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';