+++ packages/moralis/src/solUtils/index.ts
@@ -1,1 +1,1 @@
-export * from '@moralisweb3/common-sol-utils';
+export * from '@moralisweb3/sol-utils';