+++ packages/solApi/integration/setup.ts
@@ -1,6 +1,6 @@
 import { MoralisApiUtils } from '@moralisweb3/api-utils';
 import { MoralisCore } from '@moralisweb3/core';
-import { MoralisSolUtils } from '@moralisweb3/common-sol-utils';
+import { MoralisSolUtils } from '@moralisweb3/sol-utils';
 import { MoralisSolApi } from '../src/MoralisSolApi';
 import { MOCK_API_KEY } from './mocks/config';
 import { mockServer } from './mocks/mockServer';