+++ packages/solApi/src/resolvers/token/getTokenPrice.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { Camelize } from '@moralisweb3/core';
-import { SolAddress, SolNetworkish, SolAddressish, SolNative } from '@moralisweb3/common-sol-utils';
+import { SolAddress, SolNetworkish, SolAddressish, SolNative } from '@moralisweb3/sol-utils';
 import { operations } from '../../generated/types';
 import { SolNetworkResolver } from '../SolNetworkResolver';