+++ packages/streams/integration/setup.ts
@@ -1,5 +1,5 @@
 import { MoralisApiUtils } from '@moralisweb3/api-utils';
-import { MoralisEvmUtils } from '@moralisweb3/common-evm-utils';
+import { MoralisEvmUtils } from '@moralisweb3/evm-utils';
 import { MoralisCore, MoralisCoreProvider } from '@moralisweb3/core';
 import { MoralisStreams } from '../src/MoralisStreams';
 import { MOCK_API_KEY } from './mocks/config';